requirements.txt
# Core numerical + visualization stack
numpy>=1.24
scipy>=1.11
h5py>=3.9
matplotlib>=3.8
PyQt6>=6.6
PyOpenGL>=3.1.6
vispy>=0.13
pyvista>=0.43
pyvistaqt>=0.11

# Optional acceleration / neural rendering extras
numba>=0.58
cupy-cuda12x>=12.0  # pick the wheel matching your CUDA runtime
torch>=2.2
tinycudann>=1.7
nnpycgal>=0.2
